fastapi_hw/pyproject.toml
[tool.poetry]
name = "fastapi-hw"
version = "0.1.0"
description = ""
authors = ["Alex <alexsa1995@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
uvicorn = {extras = ["standard"], version = "^0.34.0"}
fastapi = "^0.115.7"
sqlalchemy = "^2.0.37"
libgravatar = "^1.0.4"
alembic = "^1.14.1"
asyncpg = "^0.30.0"
pydantic = {extras = ["email"], version = "^2.10.6"}
python-multipart = "^0.0.20"
bcrypt = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
